requirements.txt
discord.py == 2.3.2
requests == 2.31.0
pandas == 2.2.1
openai == 1.13.3